HRTList Plot corrected, avHRT line above others
added "add" argument to HRT plot (compare plot.function), so avHRt is
plotted, all other HRTs and the at last avHRT is laid over with "lines"
again.
Legend, axes etc. are plotted twice since it is necessary for plotting a
single HRT (add=FALSE) to also plot them. Additionally the legend would
be painted over by the lines if it was not plotted a second time.