adds RMSSD and moves nRMSSD into avHRT
When recalculating avHRT via calcAvHRT nRMSSD could not be recalculated
because all intervals of the original data are needed. This caused avHRT
to be different when calculated via vectorToHRT and calcAvHRT.
All intervals are only needed to calculate RMSSD. The changes separate
the calculation of RMSSD and normalising so that the latter can be done
in calcAvHRT itself every time it is called.